Revert "<ToggleButton />、<SingleDatePicker />、<Textarea />の追加 (#118)"
This reverts commit 925e8db149552cad2ba5bd06fcc0e0f2d8c34f74.
