+++ examples/javascript/App.jsx
@@ -23,7 +23,7 @@ import {
   Tooltip,
   Select,
   DropdownButton,
-  DateRangePicker,
+  DatePicker,
 } from "ingred-ui";
 import { GlobalStyle } from "./styles/globalStyle";
 
@@ -171,9 +171,9 @@ const componentList = [
         content: <RadioButton checked={true} />,
       },
       {
-        title: "DateRangePicker",
+        title: "DatePicker",
         content: (
-          <DateRangePicker
+          <DatePicker
             startDate={moment()}
             endDate={moment()}
             onDatesChange={() => {}}
