+++ src/components/DatePicker/__tests__/DatePicker.test.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import "@testing-library/jest-dom/extend-expect";
 import { cleanup } from "@testing-library/react";
-import DatePicker from "..";
+import DatePicker from "../";
 import { renderWithThemeProvider } from "../../../utils/renderWithThemeProvider";
 
 describe("DatePicker component testing", () => {
@@ -9,7 +9,7 @@ describe("DatePicker component testing", () => {
 
   test("DatePicker", () => {
     const { asFragment } = renderWithThemeProvider(
-      <DatePicker date={null} onDateChange={jest.fn()} />,
+      <DatePicker startDate={null} endDate={null} onDatesChange={jest.fn()} />,
     );
     expect(asFragment()).toMatchSnapshot();
   });
