+++ src/components/Input/Input.stories.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import styled from "styled-components";
 import Input from "./Input";
 import Typography from "../Typography";
+import { useTheme } from "../../themes/useTheme";
 import Spacer from "../Spacer";
 
 const Container = styled.div`
@@ -32,78 +33,71 @@ export default {
   },
 };
 
-export const Overview = () => (
-  <Container>
-    <RowContainer>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Empty
-        </Typography>
-        <Spacer pt={2} />
-        <Input placeholder="Place holder" />
-      </Column>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Focus
-        </Typography>
-        <Spacer pt={2} />
-        <Input />
-      </Column>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Typed
-        </Typography>
-        <Spacer pt={2} />
-        <Input value="Textfield" readOnly={true} />
-      </Column>
-    </RowContainer>
-    <RowContainer>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Disabled
-        </Typography>
-        <Spacer pt={2} />
-        <Input disabled={true} />
-      </Column>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Disabled &amp; Typed
-        </Typography>
-        <Spacer pt={2} />
-        <Input value="Textfield" readOnly={true} disabled={true} />
-      </Column>
-    </RowContainer>
-    <RowContainer>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Error
-        </Typography>
-        <Spacer pt={2} />
-        <Input error={true} readOnly={true} />
-      </Column>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Error &amp; Typed
-        </Typography>
-        <Spacer pt={2} />
-        <Input error={true} value="Textfield" readOnly={true} />
-      </Column>
-    </RowContainer>
-    <RowContainer>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Textarea
-        </Typography>
-        <Spacer pt={2} />
-        <Input multiline={true} placeholder="プレースホルダー" />
-      </Column>
-      <Column>
-        <Typography weight="bold" size="xxl">
-          Textarea(Error)
-        </Typography>
-        <Spacer pt={2} />
-        <Input error={true} multiline={true} placeholder="プレースホルダー" />
-      </Column>
-    </RowContainer>
-  </Container>
-);
+export const Overview = () => {
+  const theme = useTheme();
+  return (
+    <Container>
+      <RowContainer>
+        <Column>
+          <Typography weight="bold" size="xxl">
+            Empty
+          </Typography>
+          <Spacer pt={2} />
+          <Input placeholder="Place holder" />
+        </Column>
+        <Column>
+          <Typography weight="bold" size="xxl">
+            Focus
+          </Typography>
+          <Spacer pt={2} />
+          <Input />
+        </Column>
+      </RowContainer>
+      <RowContainer>
+        <Column>
+          <Typography weight="bold" size="xxl">
+            Typed
+          </Typography>
+          <Spacer pt={2} />
+          <Input value="Textfield" readOnly={true} />
+        </Column>
+        <Column>
+          <Typography weight="bold" size="xxl">
+            Disabled
+          </Typography>
+          <Spacer pt={2} />
+          <Input disabled={true} />
+        </Column>
+        <Column>
+          <Typography weight="bold" size="xxl">
+            Typed &amp; Disabled
+          </Typography>
+          <Spacer pt={2} />
+          <Input value="Textfield" readOnly={true} disabled={true} />
+        </Column>
+      </RowContainer>
+      <RowContainer>
+        <Column>
+          <Typography weight="bold" size="xxl">
+            Password
+          </Typography>
+          <Typography color={theme.palette.text.hint}>
+            Last
+            passを利用している方は下キーを押すとパスワードの候補が表示されます
+          </Typography>
+          <Spacer pt={2} />
+          <Input value="secure text" readOnly={true} type="password" />
+        </Column>
+      </RowContainer>
+      <RowContainer>
+        <Column>
+          <Typography weight="bold" size="xxl">
+            with Icon
+          </Typography>
+          <Spacer pt={2} />
+          <Input placeholder="検索" icon="search" />
+        </Column>
+      </RowContainer>
+    </Container>
+  );
+};
