+++ src/components/Input/Input.tsx
@@ -1,25 +1,68 @@
 import * as React from "react";
 import * as Styled from "./styled";
+import Icon from "../Icon";
+import { IconName } from "../Icon/Icon";
+import { useTheme } from "../../themes";
 
-export type Props = (
-  | React.ComponentPropsWithoutRef<"input">
-  | React.ComponentPropsWithoutRef<"textarea">
-) & {
+export type Props = React.ComponentPropsWithoutRef<"input"> & {
+  name?: string;
+  placeholder?: string;
+  type?: string;
+  value?: string;
+  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  inputRef?: React.Ref<HTMLInputElement>;
   error?: boolean;
-  multiline?: boolean;
-  resize?: "none" | "both" | "horizontal" | "vertical" | "inherit";
+  defaultValue?: string;
+  disabled?: boolean;
+  icon?: IconName;
 };
-
-const Input = React.forwardRef<HTMLTextAreaElement | HTMLInputElement, Props>(
-  ({ error = false, multiline = false, resize = "both", ...rest }, ref) => (
-    <Styled.Input
-      {...rest}
-      ref={ref as any}
-      as={multiline ? "textarea" : "input"}
-      isError={error}
-      resize={resize}
-    />
-  ),
-);
+const Input: React.FunctionComponent<Props> = ({
+  inputRef,
+  error,
+  type = "text",
+  disabled,
+  icon,
+  ...rest
+}) => {
+  const [show, setShow] = React.useState(false);
+  const theme = useTheme();
+  const onHandleToggleShowPassword = () => {
+    setShow(!show);
+  };
+  return (
+    <Styled.Container
+      className={[error ? "is-error" : "", disabled ? "is-disabled" : ""].join(
+        " ",
+      )}
+    >
+      {icon != null && (
+        <Styled.LeftIconContainer onClick={onHandleToggleShowPassword}>
+          <Icon name={icon} size="md" color={theme.palette.gray.main} />
+        </Styled.LeftIconContainer>
+      )}
+      <Styled.Input
+        {...rest}
+        ref={inputRef}
+        disabled={disabled}
+        type={type === "password" && !show ? "password" : type}
+        className={[
+          error ? "is-error" : "",
+          disabled ? "is-disabled" : "",
+        ].join(" ")}
+        hasLeftIcon={icon != null}
+        hasRightIcon={type === "password"}
+      />
+      {type === "password" && (
+        <Styled.RightIconContainer onClick={onHandleToggleShowPassword}>
+          <Icon
+            name={show ? "eye" : "eye_off"}
+            size="md"
+            color={theme.palette.black}
+          />
+        </Styled.RightIconContainer>
+      )}
+    </Styled.Container>
+  );
+};
 
 export default Input;
