+++ src/components/Input/__tests__/Input.test.tsx
@@ -11,19 +11,4 @@ describe("Input component testing", () => {
     const { asFragment } = renderWithThemeProvider(<Input />);
     expect(asFragment()).toMatchSnapshot();
   });
-
-  test("Input disabled", () => {
-    const { asFragment } = renderWithThemeProvider(<Input disabled />);
-    expect(asFragment()).toMatchSnapshot();
-  });
-
-  test("Input error", () => {
-    const { asFragment } = renderWithThemeProvider(<Input error />);
-    expect(asFragment()).toMatchSnapshot();
-  });
-
-  test("Input multiline", () => {
-    const { asFragment } = renderWithThemeProvider(<Input multiline />);
-    expect(asFragment()).toMatchSnapshot();
-  });
 });
