+++ src/components/Input/styled.ts
@@ -1,33 +1,33 @@
 import styled from "styled-components";
 import { Size, Radius } from "../../styles";
+import { colors } from "../../styles/color";
+
+export const Container = styled.div`
+  position: relative;
+  display: flex;
+  align-items: center;
+  width: 100%;
+`;
 
 export const Input = styled.input<{
-  isError: boolean;
-  width?: string | number;
-  resize?: string;
+  hasLeftIcon: boolean;
+  hasRightIcon: boolean;
 }>`
-  ${({ width }) =>
-    width ? `width: ${isNaN(+width) ? width : width + "px"}` : ""};
-  padding: 10px 8px;
+  display: inline-flex;
+  padding: 10px ${({ hasRightIcon }) => (hasRightIcon ? "32px" : "8px")} 10px
+    ${({ hasLeftIcon }) => (hasLeftIcon ? "35px" : "8px")};
+  flex: 1;
   border: 0;
   line-height: 20px;
   font-size: 14px;
-  color: ${({ theme, isError }) =>
-    isError ? theme.palette.danger.main : theme.palette.black};
-  background-color: ${({ theme, isError }) =>
-    isError
-      ? theme.palette.danger.highlight
-      : theme.palette.background.default};
+  color: ${({ theme }) => theme.palette.black};
+  background-color: transparent;
   border: ${Size.Border.Small} solid ${({ theme }) => theme.palette.divider};
   border-radius: ${Radius.SMALL};
   box-shadow: 0 ${Size.Border.Normal} 0 0
-    ${({ theme, isError }) =>
-      isError ? theme.palette.danger.highlight : theme.palette.gray.light}
-    inset;
-  border-color: ${({ theme, isError }) =>
-    isError ? theme.palette.danger.main : theme.palette.divider};
+    ${({ theme }) => theme.palette.gray.light} inset;
+  background-color: ${({ theme }) => theme.palette.background.default};
   overflow: hidden;
-  resize: ${({ resize }) => resize};
   /* lastpassのicon用 */
   background-position: calc(100% - 35px) 50% !important;
   &:focus {
@@ -37,14 +37,40 @@ export const Input = styled.input<{
   &::placeholder {
     color: ${({ theme }) => theme.palette.text.hint};
   }
+  /* IE */
+  input:-ms-input-placeholder {
+    color: ${({ theme }) => theme.palette.text.hint};
+  }
   /* Edge */
   &::-ms-input-placeholder {
     color: ${({ theme }) => theme.palette.text.hint};
   }
-  &:disabled {
+  &.is-disabled {
     color: ${({ theme }) => theme.palette.text.disabled};
     border-color: ${({ theme }) => theme.palette.divider};
     box-shadow: none;
     background-color: ${({ theme }) => theme.palette.gray.light};
   }
+  &.is-error {
+    color: ${({ theme }) => theme.palette.danger.main};
+    border-color: ${({ theme }) => theme.palette.danger.main};
+    box-shadow: 0 ${Size.Border.Normal} 0 0
+      ${({ theme }) => theme.palette.danger.highlight} inset;
+    background-color: ${colors.red[100]};
+  }
+`;
+
+const IconContainer = styled.div`
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  cursor: pointer;
+`;
+
+export const LeftIconContainer = styled(IconContainer)`
+  left: 10px;
+`;
+
+export const RightIconContainer = styled(IconContainer)`
+  right: 10px;
 `;
