+++ src/components/TextField/TextField.stories.tsx
@@ -49,18 +49,6 @@ export const Overview = () => {
             errorText={"数値で入力してください"}
           />
         </Column>
-        <Column>
-          <Typography weight="bold" size="xxl">
-            Error(Textarea)
-          </Typography>
-          <Spacer pt={2} />
-          <TextField
-            value="Textfield"
-            readOnly={true}
-            errorText={"入力が正しくありません"}
-            multiline={true}
-          />
-        </Column>
       </RowContainer>
       <RowContainer>
         <Column>
@@ -73,15 +61,6 @@ export const Overview = () => {
           </Typography>
           <Spacer pt={2} />
           <TextField value="secure text" readOnly={true} type="password" />
-        </Column>
-      </RowContainer>
-      <RowContainer>
-        <Column>
-          <Typography weight="bold" size="xxl">
-            With icon
-          </Typography>
-          <Spacer pt={2} />
-          <TextField placeholder="検索" icon="search" />
         </Column>
       </RowContainer>
     </Container>
