--- src/components/ToggleButton/__tests__/ToggleButton.test.tsx
@@ -1,14 +1,0 @@
-import * as React from "react";
-import "@testing-library/jest-dom/extend-expect";
-import { cleanup } from "@testing-library/react";
-import ToggleButton from "..";
-import { renderWithThemeProvider } from "../../../utils/renderWithThemeProvider";
-
-describe("ToggleButton component testing", () => {
-  afterEach(cleanup);
-
-  test("ToggleButton", () => {
-    const { asFragment } = renderWithThemeProvider(<ToggleButton />);
-    expect(asFragment()).toMatchSnapshot();
-  });
-});
