+++ src/components/index.ts
@@ -32,9 +32,6 @@ export * from "./DataTable";
 export { default as DatePicker } from "./DatePicker";
 export * from "./DatePicker";
 
-export { default as DateRangePicker } from "./DateRangePicker";
-export * from "./DateRangePicker";
-
 export { default as Divider } from "./Divider";
 export * from "./Divider";
 
@@ -112,9 +109,6 @@ export * from "./TextField";
 
 export { default as Toast } from "./Toast";
 export * from "./Toast";
-
-export { default as ToggleButton } from "./ToggleButton";
-export * from "./ToggleButton";
 
 export { default as Tooltip } from "./Tooltip";
 export * from "./Tooltip";
